services/rag-anything/requirements.txt
# Core dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
structlog==24.4.0
pydantic==2.9.0
pydantic-settings==2.6.0

# Document parsers (per Story 0.1 spike recommendation)
pypdf>=4.0.0
python-docx>=1.0.0
python-pptx>=1.0.0
pandas>=2.0.0

# Image processing
Pillow>=10.0.0

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
httpx==0.27.0
pytest-cov==5.0.0
python-multipart==0.0.20
